Make sure the correct PIDs are stored
Previously the PID of the logging subprocess was being stored instead of
the PID of the program specified by the Procfile.

This change means the correct PID is now stored, but it also means we
sadly loose prefixed logging :cry: - if anyone has any suggestions on
ways to get it back then I'd be pleased to hear them!

Fixes #29